Fixed sample for duration and added sample for codec_name
I've updated the README file, because the existing example for getting the duration of a file was not working for file-types like WebM (see #103).

To not to loose the sample that was already in place, I changed it to return the codec-name for the video. Maybe returning width or height of a video is more used, but we can't have a sample for all ;)